Complete set of routines to interact with the GIS database
Implemented:
- Counting features by type
- Clearing database of all features by type
- Fetching the coordinates of a point
- Fetching all features inside a circle
- Calculating distance from feature to point
- Checking for line-of-sight between two points
- Checking whether a location is obstructed
- Adding new points to the database
- Updating existing points in the database

Also, we now store coordinate pairs as Doubles, for convenience.

+++ src/obj.swift
@@ -6,12 +6,12 @@ import Foundation
 
 struct FCDVehicle {
 	let id: UInt
-	let xgeo: Float
-	let ygeo: Float
-	let speed: Float
+	let xgeo: Double
+	let ygeo: Double
+	let speed: Double
 }
 
 struct FCDTimestep {
-	let time: Float
+	let time: Double
 	let vehicles: [FCDVehicle]
 }
